Consolidate Gyro helper functions
