+++ src/pools/gyro3Pool/constants.ts
@@ -5,18 +5,6 @@ import { BigNumber, parseFixed } from '@ethersproject/bignumber';
 export const _MAX_IN_RATIO: BigNumber = parseFixed('0.3', 18);
 export const _MAX_OUT_RATIO: BigNumber = parseFixed('0.3', 18);
 
-// SQRT constants
-
-export const SQRT_1E_NEG_1 = BigNumber.from('316227766016837933');
-export const SQRT_1E_NEG_3 = BigNumber.from('31622776601683793');
-export const SQRT_1E_NEG_5 = BigNumber.from('3162277660168379');
-export const SQRT_1E_NEG_7 = BigNumber.from('316227766016837');
-export const SQRT_1E_NEG_9 = BigNumber.from('31622776601683');
-export const SQRT_1E_NEG_11 = BigNumber.from('3162277660168');
-export const SQRT_1E_NEG_13 = BigNumber.from('316227766016');
-export const SQRT_1E_NEG_15 = BigNumber.from('31622776601');
-export const SQRT_1E_NEG_17 = BigNumber.from('3162277660');
-
 // POW3 constant
 // Threshold of x where the normal method of computing x^3 would overflow and we need a workaround.
 // Equal to 4.87e13 scaled; 4.87e13 is the point x where x**3 * 10**36 = (x**2 native) * (x native) ~ 2**256
